Optimize object lookup when dictionary is sorted
Check if the OD is sorted and then use a binary search for the lookup,
otherwise fall back to the regular linear search.

Other algorithms can be added in the same way to use hash tables or
other methods to speed up the lookup even if the OD is not pre-sorted.

Note that PDO accesses bypass the lookup entirely, so this optimization
is mostly for cases where a large dictionary is frequently accessed by
other means.

Change-Id: Ic1ff6aadfa1be3f63720f12ee1823f4f41fd1d88